requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
